Migrate realm_finish_reencryption_maintenance tests
