+++ libparsec/crates/protocol/tests/authenticated_cmds/v4/realm_finish_reencryption_maintenance.rs
@@ -1,35 +1,27 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
+#![allow(clippy::duplicate_mod)]
+
+// The compat module allows to re-use tests from previous major API
+#[path = "../v3/realm_finish_reencryption_maintenance.rs"]
+mod compat;
+
 use super::authenticated_cmds;
 
 // Request
 
-pub fn req() {
-    // TODO #4545: Implement test
-}
+pub use compat::req;
 
 // Responses
 
-pub fn rep_ok() {
-    // TODO #4545: Implement test
-}
+pub use compat::rep_ok;
 
-pub fn rep_not_allowed() {
-    // TODO #4545: Implement test
-}
+pub use compat::rep_not_allowed;
 
-pub fn rep_not_found() {
-    // TODO #4545: Implement test
-}
+pub use compat::rep_not_found;
 
-pub fn rep_bad_encryption_revision() {
-    // TODO #4545: Implement test
-}
+pub use compat::rep_bad_encryption_revision;
 
-pub fn rep_not_in_maintenance() {
-    // TODO #4545: Implement test
-}
+pub use compat::rep_not_in_maintenance;
 
-pub fn rep_maintenance_error() {
-    // TODO #4545: Implement test
-}
+pub use compat::rep_maintenance_error;
